requirements.txt
openai==1.12.0
pymongo==4.6.1
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3 